requirements.txt
numpy==1.21.2
pillow>=9.4.0
scipy==1.7.3

# "Torch not compiled with CUDA enabled": https://github.com/pytorch/pytorch/issues/30664#issuecomment-757431613
# pip install torch -f https://download.pytorch.org/whl/torch_stable.html
torch==2.0.0

torchvision==0.15.1
flask==2.0.2
rdkit-pypi==2022.03.3    # colorama==0.4.6